third_party/blink/web_tests/external/wpt/storage-access-api/storage-access-permission.sub.https.window-expected.txt

This is a testharness.js-based test.
PASS Permissions grants are observable across same-origin iframes
FAIL Permission default state can be queried assert_equals: expected "prompt" but got "denied"
PASS Permission granted state can be queried
PASS Permission denied state can be queried
PASS Permission state can be observed
Harness: the test ran to completion.
